Earnings-Intel/requirements.txt
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
yfinance==0.2.54
python-dotenv==1.0.1
vaderSentiment==3.3.2
scikit-learn==1.6.1
statsmodels==0.14.4
pyarrow==18.1.0
sentence-transformers==3.4.1
fastapi==0.115.8
uvicorn==0.34.0
joblib==1.4.2
